main view controller, major refactor on gaming and racing controllers
